+++ src/global/listingLimit.ts
@@ -9,3 +9,5 @@ export const LEADERBOARD_LISTING_LIMIT: number = 10;
 export const LATEST_POSTS_LIMIT: number = 8;
 
 export const VOTES_LISTING_LIMIT: number = 10;
+
+export const BOUNTIES_LISTING_LIMIT = 10;
